streamlit_app/requirements.txt
streamlit
requests
pandas
streamlit-audiorecorder
